Refactored Social icons into component.
